test_langgraph.py
try:
    from langgraph.prebuilt import create_react_agent
    print("Found create_react_agent in langgraph.prebuilt")
except ImportError:
    print("NOT found in langgraph.prebuilt")

try:
    from langgraph.prebuilt import chat_agent_executor
    print("Found chat_agent_executor in langgraph.prebuilt")
except ImportError:
    print("NOT found chat_agent_executor")
