requirements.txt
langchain
langchain-community
langchain-google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
notion-client
python-dotenv
pydantic
pydantic-settings
schedule
